Implement Episode Grouper script
Jellyfin doesn't automatically group duplicate/multi-version Series
episodes. This is annoying if you have alot of series and for each
episode have rendered multipler versions (HD+mobile for example).
It is not viable to manually merge the episodes.

This script will merge episodes in a season that has the
same imdbId, tvdb id or name (excluding any version flag at the end
of the filename).